+++ src/renderer/app.component.tsx
@@ -1,6 +1,6 @@
 import { MemoryRouter as Router, Routes, Route } from 'react-router-dom';
 import './app.component.css';
-import ParanextDockLayout from '@renderer/components/docking/paranext-dock-layout.component';
+import PlatformDockLayout from '@renderer/components/docking/platform-dock-layout.component';
 import TestContext from '@renderer/context/papi-context/test.context';
 import PlatformBibleToolbar from './components/platform-bible-toolbar';
 
@@ -8,7 +8,7 @@ function Main() {
   return (
     <TestContext.Provider value="test">
       <PlatformBibleToolbar />
-      <ParanextDockLayout />
+      <PlatformDockLayout />
     </TestContext.Provider>
   );
 }
