+++ src/renderer/components/dialogs/dialog-definition.model.ts
