+++ src/renderer/components/docking/platform-dock-layout.component.css
@@ -3,7 +3,7 @@
   margin: 4px;
 }
 
-.dock-layout .dock-panel.dock-style-paranext {
+.dock-layout .dock-panel.dock-style-platform-bible {
   border: 0;
 }
 
@@ -21,7 +21,7 @@
   border: 0;
 }
 
-.dock-layout .dock-panel.dock-style-paranext .dock-tab {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-tab {
   margin-right: 0;
   border: 1px solid #5c5c5c;
   background: #bfbfbf;
@@ -29,7 +29,7 @@
   flex: 1 0 auto;
 }
 
-.dock-layout .dock-panel.dock-style-paranext .dock-tab > div {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-tab > div {
   padding-left: 7px;
 }
 
@@ -37,47 +37,47 @@
  * Show borders on the sides of the dock-nav, not on the very edges of the dock-tabs,
  * so the borders show even when the tabs don't fit in the space and you have to scroll
  */
-.dock-layout .dock-panel.dock-style-paranext .dock-tab:first-child {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-tab:first-child {
   border-left: 0;
 }
 
-.dock-layout .dock-panel.dock-style-paranext .dock-tab:nth-last-child(2) {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-tab:nth-last-child(2) {
   border-right: 0;
 }
 
-.dock-layout .dock-panel.dock-style-paranext .dock-nav {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-nav {
   border-left: 1px solid #5c5c5c;
   border-right: 1px solid #5c5c5c;
 }
 
-.dock-layout .dock-panel.dock-style-paranext .dock-tab.dock-tab-active {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-tab.dock-tab-active {
   background: #a6c9ff;
   color: #191919;
   border-bottom: 1px solid #a6c9ff;
 }
 
 /* Curved corners including on floating window */
-.dock-layout .dock-panel.dock-style-paranext,
-.dock-layout .dock-panel.dock-style-paranext .dock-bar,
-.dock-layout .dock-panel.dock-style-paranext .dock-nav,
-.dock-layout .dock-panel.dock-style-paranext .dock-nav-wrap,
-.dock-layout .dock-panel.dock-style-paranext .dock-tab {
+.dock-layout .dock-panel.dock-style-platform-bible,
+.dock-layout .dock-panel.dock-style-platform-bible .dock-bar,
+.dock-layout .dock-panel.dock-style-platform-bible .dock-nav,
+.dock-layout .dock-panel.dock-style-platform-bible .dock-nav-wrap,
+.dock-layout .dock-panel.dock-style-platform-bible .dock-tab {
   border-radius: 5px 5px 0 0;
 }
 
-.dock-layout .dock-panel.dock-style-paranext .dock-bar {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-bar {
   background: #8c8c8c;
   padding: 0;
   border: 0;
 }
 
-.dock-layout .dock-panel.dock-style-paranext .dock-tab-hit-area {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-tab-hit-area {
   /* expand the invisible dock-tab-hit-area a bit to cover the border area above it */
   left: -1px;
   right: -1px;
 }
 
-.dock-layout .dock-panel.dock-style-paranext .dock-nav-wrap {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-nav-wrap {
   /*
    * Remove the handle bar at the top of each dock-tab. Note: Now, the tab
    * group can only be grabbed by grabbing the intersection between two tabs
@@ -91,7 +91,7 @@
 }
 
 /* tabs take all the space */
-.dock-layout .dock-panel.dock-style-paranext .dock-nav-list {
+.dock-layout .dock-panel.dock-style-platform-bible .dock-nav-list {
   flex-grow: 1;
 }
