+++ src/renderer/components/docking/platform-dock-layout.component.test.ts
@@ -22,7 +22,7 @@ import {
   addWebViewToDock,
   getFloatPosition,
   loadTab,
-} from './paranext-dock-layout.component';
+} from './platform-dock-layout.component';
 
 describe('Dock Layout Component', () => {
   const mockDockLayout = mock(DockLayout);
