+++ src/renderer/components/docking/platform-panel.component.css
