+++ src/renderer/components/docking/platform-tab-title.component.css
