+++ src/renderer/components/projects/download-update-project-tab.component.scss
