+++ src/renderer/components/projects/open-multiple-projects-tab.component.scss
