+++ src/renderer/components/projects/project-list.component.tsx
