+++ src/renderer/hooks/papi-hooks/use-dialog-callback.hook.ts
@@ -1,4 +1,4 @@
-import { DialogTabTypes, DialogTypes } from '@renderer/components/dialogs/dialog.data';
+import { DialogTabTypes, DialogTypes } from '@renderer/components/dialogs/dialog-definition.model';
 import dialogService from '@shared/services/dialog.service';
 import { getErrorMessage } from '@shared/utils/util';
 import { useCallback, useEffect, useRef, useState } from 'react';
