+++ src/renderer/services/papi-frontend.service.ts
@@ -12,7 +12,7 @@ import { papiNetworkService, PapiNetworkService } from '@shared/services/network
 import { papiWebViewService, PapiWebViewService } from '@shared/services/web-view.service';
 import internetService, { InternetService } from '@shared/services/internet.service';
 import dataProviderService, { DataProviderService } from '@shared/services/data-provider.service';
-import { ProjectLookupServiceType } from '@shared/services/project-lookup.service.model';
+import { ProjectLookupServiceType } from '@shared/services/project-lookup.service-model';
 import projectLookupService from '@shared/services/project-lookup.service';
 import {
   papiFrontendProjectDataProviderService,
@@ -22,7 +22,7 @@ import papiContext, { PapiContext } from '@renderer/context/papi-context';
 import papiHooks, { PapiHooks } from '@renderer/hooks/papi-hooks';
 import settingsService, { SettingsService } from '@shared/services/settings.service';
 import dialogService from '@shared/services/dialog.service';
-import { DialogService } from '@shared/services/dialog.service.model';
+import { DialogService } from '@shared/services/dialog.service-model';
 
 // IMPORTANT NOTES:
 // 1) When adding new services here, consider whether they also belong in papi-backend.service.ts.
